Created LanePropertiesContextManager
